fix: add user validations

+++ apps/backend/src/main/java/org/backend/core/room/presentation/RoomController.java
@@ -1,12 +1,14 @@
 package org.backend.core.room.presentation;
 
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.Validator;
 import org.backend.config.security.SessionUser;
 import org.backend.core.room.application.RoomServiceInterface;
 import org.backend.core.room.application.dto.CreateRoomDto;
 import org.backend.core.room.application.dto.DeleteRoomDto;
 import org.backend.core.room.application.dto.GetRoomsDto;
-import org.backend.core.room.presentation.mapper.RoomResponseDto;
 import org.backend.core.room.presentation.mapper.RoomMapper;
+import org.backend.core.room.presentation.mapper.RoomResponseDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
@@ -19,14 +21,25 @@ import java.util.UUID;
 public class RoomController {
     private final RoomServiceInterface roomService;
 
-    public RoomController(RoomServiceInterface roomService) {
+    private final Validator validator;
+
+    public RoomController(
+            RoomServiceInterface roomService,
+            Validator validator
+    ) {
         this.roomService = roomService;
+        this.validator = validator;
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public void create(@RequestBody CreateRoomDto dto, @AuthenticationPrincipal SessionUser principal) {
         dto.addUserId(principal.id());
+
+        var validations = validator.validate(dto);
+        if (!validations.isEmpty()) {
+            throw new ConstraintViolationException(validations);
+        }
 
         roomService.create(dto);
     }
@@ -37,6 +50,11 @@ public class RoomController {
         dto.setRoomId(id);
         dto.setUserId(principal.id());
 
+        var validations = validator.validate(dto);
+        if (!validations.isEmpty()) {
+            throw new ConstraintViolationException(validations);
+        }
+
         roomService.delete(dto);
     }
 
@@ -44,6 +62,11 @@ public class RoomController {
     public List<RoomResponseDto> getRooms(@AuthenticationPrincipal SessionUser principal) {
         GetRoomsDto dto = new GetRoomsDto();
         dto.setUserId(principal.id());
+
+        var validations = validator.validate(dto);
+        if (!validations.isEmpty()) {
+            throw new ConstraintViolationException(validations);
+        }
 
         return roomService.get(dto).stream()
                 .map(RoomMapper::toResponseDto)
